Split KnowledgeService helpers into focused modules (412 → 304 lines)
The class itself stays in KnowledgeService.ts — buildContext is one
cohesive orchestration flow that benefits from being read top-to-
bottom. The supporting concerns lift out into ./knowledge-service/:

- types.ts        request/response shapes + KnowledgeLane union +
                  CORE_PRIORITY_KEYS (with comment explaining why
                  it's a curated subset) + PERSONAL_MEMORY_TYPES +
                  LANE_DIRECTIVES (per-lane preamble strings)
- scoring.ts      normalizeText / extractKeywords / scoreText /
                  scoreProjectMemory / scoreScratchpadMemory /
                  safeExcerpt — pure text math with comments on
                  the weight choices (why personal-type gets +5,
                  why ≥3-char keyword filter, why 12-keyword cap)
- formatting.ts   parseCoreValue / formatCoreMemory /
                  formatRetrievedMemory / dedupeRetrievedMemory
- sources.ts      loadRulesetMemory — the four-YAML disk loader,
                  silent-skip-on-error explained

Class file is still 304 lines but every reader-hostile chunk
(85-line constants block, two-tier scoring math, format helpers)
is now a one-import away instead of inline.
